[core] Move the scanf scan into code_scan.py on the loader's component list and trim the tests
