+++ esphome/writer.py
@@ -246,6 +246,9 @@ def copy_src_tree():
         Path(
             "esphome/core/ring_buffer.h"
         ),  # moved to components/ring_buffer/, removed in 2026.11.0
+        Path(
+            "esphome/core/sscanf_no_float.h"
+        ),  # only the esp32 sscanf wrap includes it
     }
     include_l = []
     for target, _ in source_files_l:
