--- tests/component_tests/esp32/config/sscanf_stub_ble_c2.yaml
@@ -1,9 +1,0 @@
-esphome:
-  name: test
-
-esp32:
-  variant: esp32c2
-  framework:
-    type: esp-idf
-
-esp32_ble_tracker:
