--- tests/component_tests/esp32/config/sscanf_stub_ble_lambda_forced.yaml
@@ -1,19 +1,0 @@
-esphome:
-  name: test
-
-esp32:
-  board: esp32dev
-  framework:
-    type: esp-idf
-    advanced:
-      enable_full_scanf: false
-
-esp32_ble_tracker:
-
-sensor:
-  - platform: template
-    name: Parsed
-    lambda: |-
-      float value = 0;
-      sscanf("1.5", "%f", &value);
-      return value;
