+++ tests/component_tests/esp32/test_esp32.py
@@ -1443,13 +1443,6 @@ def test_mbedtls_tls_openthread_requires_server_and_extras(
             "sscanf_stub_ble_lambda.yaml", "sscanf", False, id="sscanf_lambda_float"
         ),
         pytest.param(
-            "sscanf_stub_ble_lambda_forced.yaml",
-            "sscanf",
-            True,
-            id="sscanf_lambda_float_forced",
-        ),
-        pytest.param("sscanf_stub_ble_c2.yaml", "sscanf", False, id="sscanf_c2_rom"),
-        pytest.param(
             "sscanf_stub_ble_full_printf.yaml", "sscanf", True, id="sscanf_full_printf"
         ),
         pytest.param("exclusion_reincludes.yaml", "sscanf", False, id="sscanf_no_ble"),
