+++ tests/unit_tests/test_config_helpers.py
@@ -2,26 +2,18 @@
 
 from collections.abc import Callable
 from pathlib import Path
-from unittest.mock import MagicMock, patch
+from unittest.mock import patch
 
 import pytest
 
 from esphome.config_helpers import (
-    external_components_use_scanf_float,
     filter_source_files_from_defines,
     filter_source_files_from_platform,
     frameworks_for_platforms,
     get_logger_level,
-    includes_use_scanf_float,
     iter_include_files,
-    keep_float_scanf,
-    lambdas_use_scanf_float,
-    user_code_uses_scanf_float,
 )
 from esphome.const import (
-    CONF_ESPHOME,
-    CONF_INCLUDES,
-    CONF_INCLUDES_C,
     CONF_LEVEL,
     CONF_LOGGER,
     KEY_CORE,
@@ -29,9 +21,7 @@ from esphome.const import (
     KEY_TARGET_PLATFORM,
     PlatformFramework,
 )
-from esphome.core import CORE, Define, Lambda
-from esphome.loader import CORE_COMPONENTS_PATH
-from esphome.types import ConfigType
+from esphome.core import Define
 
 
 def test_filter_source_files_from_platform_esp32() -> None:
@@ -186,69 +176,6 @@ def test_filter_source_files_from_defines() -> None:
         assert sorted(filter_func()) == ["automation.cpp", "filter.cpp"]
 
 
-@pytest.mark.parametrize(
-    ("src", "expected"),
-    [
-        # Basic float formats
-        ('sscanf(buf, "%f", &v)', True),
-        ('sscanf(buf, "%F", &v)', True),
-        ('sscanf(buf, "%e", &v)', True),
-        ('sscanf(buf, "%E", &v)', True),
-        ('sscanf(buf, "%g", &v)', True),
-        ('sscanf(buf, "%G", &v)', True),
-        ('sscanf(buf, "%a", &v)', True),
-        ('sscanf(buf, "%A", &v)', True),
-        # With modifiers
-        ('sscanf(buf, "%lf", &v)', True),
-        ('sscanf(buf, "%Lf", &v)', True),
-        ('sscanf(buf, "%8lf", &v)', True),
-        ('sscanf(buf, "%*f")', True),
-        ('sscanf(buf, "%.2f", &v)', True),
-        # Mixed formats, including a ';' inside the format string
-        ('sscanf(buf, "%d,%f", &a, &b)', True),
-        ('sscanf(buf, "%d;%f", &a, &b)', True),
-        ('sscanf(buf, "a;b", &a); sscanf(buf, "%d", &b); printf("%f", v)', False),
-        # Escaped quotes, nested calls, a ')' inside the format, an unterminated call
-        ('sscanf(buf, "\\"%f\\"", &v)', True),
-        ('sscanf(f(x), "%f", &v)', True),
-        ('sscanf(buf, "%d)", &a); g("%f")', False),
-        ('sscanf(buf, "%d", &a; g("%f")', False),
-        ('sscanf(buf, "%f"', True),
-        # fscanf and std::sscanf
-        ('fscanf(fp, "%f", &v)', True),
-        ('std::sscanf(buf, "%f", &v)', True),
-        # Multi-line
-        ('sscanf(buf,\n"%f", &v)', True),
-        # No float format
-        ('sscanf(buf, "%d", &v)', False),
-        ('sscanf(buf, "%s", s)', False),
-        # printf not scanf
-        ('printf("%f", val)', False),
-        # %f in a different statement after scanf
-        ('sscanf(buf, "%d", &x); printf("%f", val);', False),
-        # scanf %f in comment only
-        ('// sscanf(buf, "%f", &v)\nsscanf(buf, "%d", &x)', False),
-        ('/* sscanf(buf, "%f") */\nsscanf(buf, "%d", &x)', False),
-    ],
-)
-def test_lambdas_use_scanf_float(src: str, expected: bool) -> None:
-    """Test scanf float detection in lambda source."""
-    config: ConfigType = {"test": [Lambda(src)]}
-    assert lambdas_use_scanf_float(config) is expected
-
-
-def test_lambdas_use_scanf_float_no_lambdas() -> None:
-    """Test with config containing no lambdas."""
-    config: ConfigType = {"key": "value", "list": [1, 2]}
-    assert lambdas_use_scanf_float(config) is False
-
-
-def test_lambdas_use_scanf_float_nested() -> None:
-    """Test detection in deeply nested config."""
-    config: ConfigType = {"a": {"b": {"c": [Lambda('sscanf(buf, "%f", &v)')]}}}
-    assert lambdas_use_scanf_float(config) is True
-
-
 def test_iter_include_files(setup_core: Path) -> None:
     """Files yield their name, directory members keep the directory prefix, system headers are skipped."""
     (setup_core / "one.h").write_text("")
@@ -259,89 +186,3 @@ def test_iter_include_files(setup_core: Path) -> None:
         for _, basename in iter_include_files(["one.h", "lib", "<cstdio>"])
     }
     assert files == {"one.h", "lib/two.h"}
-
-
-def test_includes_use_scanf_float(setup_core: Path) -> None:
-    """A float scanf in an includes: or includes_c: file is found; comments and non-source files are ignored."""
-    (setup_core / "parse.h").write_text(
-        'static void f(const char *b) { float v; sscanf(b, "%f", &v); }'
-    )
-    (setup_core / "lib").mkdir()
-    (setup_core / "lib" / "int.h").write_text(
-        'static int g(const char *b) { int v; sscanf(b, "%d", &v); return v; }'
-    )
-    (setup_core / "lib" / "notes.txt").write_text('sscanf(b, "%f", &v)')
-    (setup_core / "commented.h").write_text(
-        '// sscanf(b, "%f", &v)\n/* sscanf(b, "%g", &v) */'
-    )
-
-    def config(
-        includes: tuple[str, ...] = (), includes_c: tuple[str, ...] = ()
-    ) -> ConfigType:
-        return {
-            CONF_ESPHOME: {
-                CONF_INCLUDES: list(includes),
-                CONF_INCLUDES_C: list(includes_c),
-            }
-        }
-
-    assert includes_use_scanf_float(config(("parse.h",))) is True
-    assert includes_use_scanf_float(config(includes_c=("parse.h",))) is True
-    assert includes_use_scanf_float(config(("lib",))) is False
-    assert includes_use_scanf_float(config(("commented.h", "<cstdio>"))) is False
-    assert includes_use_scanf_float({}) is False
-    assert user_code_uses_scanf_float(config(("parse.h",))) is True
-    assert user_code_uses_scanf_float({"test": [Lambda('sscanf(b, "%f", &v)')]}) is True
-
-
-def test_keep_float_scanf() -> None:
-    """Explicit values win; unset follows whether user code scans a float."""
-    float_config: ConfigType = {"test": [Lambda('sscanf(b, "%f", &v)')]}
-    int_config: ConfigType = {"test": [Lambda('sscanf(b, "%d", &v)')]}
-    assert keep_float_scanf(True, int_config, "x", "y") is True
-    assert keep_float_scanf(False, float_config, "x", "y") is False
-    assert keep_float_scanf(None, float_config, "x", "y") is True
-    assert keep_float_scanf(None, int_config, "x", "y") is False
-
-
-def test_external_components_use_scanf_float(setup_core: Path) -> None:
-    """Sources of components outside the ESPHome tree are scanned; in-tree ones are skipped."""
-    ext = setup_core / "ext_comp"
-    ext.mkdir()
-    (ext / "__init__.py").write_text("")
-    (ext / "sensor").mkdir()
-    (ext / "sensor" / "parse.cpp").write_text(
-        'void f(const char *b) { float v; sscanf(b, "%f", &v); }'
-    )
-    fake = MagicMock()
-    fake.module.__file__ = str(ext / "__init__.py")
-    core_component = MagicMock()
-    core_component.module.__file__ = str(
-        CORE_COMPONENTS_PATH / "sensor" / "__init__.py"
-    )
-    core_domain = MagicMock()
-    core_domain.module.__file__ = str(
-        CORE_COMPONENTS_PATH.parent / "core" / "config.py"
-    )
-    ext_int = setup_core / "ext_int"
-    ext_int.mkdir()
-    (ext_int / "__init__.py").write_text("")
-    (ext_int / "parse.cpp").write_text(
-        'int g(const char *b) { int v; sscanf(b, "%d", &v); return v; }'
-    )
-    ext_int_comp = MagicMock()
-    ext_int_comp.module.__file__ = str(ext_int / "__init__.py")
-    lookup = {
-        "ext_comp": fake,
-        "ext_int": ext_int_comp,
-        "sensor": core_component,
-        "esphome": core_domain,
-        "missing": None,
-    }
-
-    with patch("esphome.loader.get_component", side_effect=lookup.get):
-        CORE.loaded_integrations = {"sensor", "esphome", "missing", "ext_int"}
-        assert external_components_use_scanf_float() is False
-        CORE.loaded_integrations = {"sensor", "ext_comp"}
-        assert external_components_use_scanf_float() is True
-        assert user_code_uses_scanf_float({}) is True
